% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_gdppop.R
\docType{data}
\name{gw_gdppop}
\alias{gw_gdppop}
\title{GDP, Population, and GDP per Capita (Gleditsch-Ward System)}
\format{
A data frame with the following 8 variables.
\describe{
\item{\code{gwcode}}{a numeric vector for the Gleditsch-Ward state code}
\item{\code{year}}{a numeric vector for the year}
\item{\code{mrgdppc}}{a numeric vector for the estimated GDP per capita in a given year. See Details.}
\item{\code{sd_mrgdppc}}{a numeric vector for the standard deviation of estimated GDP per capita in a given year.}
\item{\code{pwtrgdp}}{a numeric vector for the estimated real GDP in a given year. See Details.}
\item{\code{sd_pwtrgdp}}{a numeric vector for the standard deviation of estimated real GDP in a given year.}
\item{\code{pwtpop}}{a numeric vector for the estimated population in a given year. See Details.}
\item{\code{sd_pwtpop}}{a numeric vector for the standard deviation of estimated population in a given year.}
}
}
\usage{
gw_gdppop
}
\description{
These are estimates of democracy for Gleditsch-Ward state system members.
}
\details{
Based on my reading of Fariss et al. (2022), I think the following information
gathered from their simulations make sense for suggested defaults. You may
want to get their actual simulations if you want something else, but I think
what's included here is good for most use cases.

For additional clarification, the suggested defaults included in this data set
are:
\itemize{
\item GDP per capita: real GDP per capita in prices constant across countries and
over time (in 2011 international dollars, PPP).
\item GDP: expenditure-side real GDP in prices constant across countries and over
time (in millions of 2017 international dollars, PPP)
\item Population: total population (in millions)
}

The GDP per capita measure is anchored around the Maddison Project Database.
The GDP and population measures are anchored around Penn World Tables (10.0).
You can create a rough estimate of GDP per capita from the Penn World Table
simulations based on the information in this data set. It's free and the cops
can't stop you.

I also honor the authors' please to include the standard deviation of these
estimates as well.
}
\references{
Please cite Miller (2022) for \pkg{peacesciencer}. Beyond that, cite the
following, contingent on which democracy estimate you are using.

Additionally, please cite Fariss et al. (2022) for the underlying data.
That citation follows.

Fariss, Christopher, J., Therese Anders, Jonathan N. Markowitz, and Miriam
Barnum. 2022. "New Estimates of Over 500 Years of Historic GDP and Population
Data." \emph{Journal of Conflict Resolution} 66(3): 553--91.

You can also cite the underlying data being estimated/simulated by Fariss et
al. (2022). For the Maddison Project Database:

Bolt, Jutta, Robert Inklaar, Herman de Jong, and Luiten Janvan Zanden. 2018.
"Rebasing 'Maddison': New Income Comparisons and the Shape of Long-Run Economic
Development." \emph{Maddison Project Working paper 10}.

For Penn World Table:

Feenstra, Robert C., Robert Inklaar, and Marcel P. Timmer. 2015. "The Next
Generation of the Penn World Table." \emph{American Economic Review} 105(10):
3150--82.
}
\keyword{datasets}
