% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_panel.R
\name{state_panel}
\alias{state_panel}
\title{Create a Panel of State-Years from the Correlates of War or Gleditsch-Ward system.}
\usage{
state_panel(system = "cow")
}
\arguments{
\item{system}{a state system (either "cow" or "gw")}
}
\value{
\code{state_panel()} returns a data frame of state years corresponding with
either the Correlates of War or the Gleditsch-Ward system.
}
\description{
\code{state_panel()} is a function to create a panel of state-years
from one of two major state systems in international relations scholarship.
}
\details{
This function leans on \code{cw_system} and \code{gw_system} in this package.
}
\examples{

head(state_panel(), 10)
head(state_panel(system='gw'), 10)

}
